requirements.txt
tensorflow>=2.17.0
keras>=3.4.1
absl-py>=2.0.0

torch
torchvision
timm
pillow
